Use jj log instead of show (#14)
jj show is equivalent to jj log except it always appends the diff to the
end of the output.

jj show was generating and appending the full diff after the template
fields, and since I had a custom diff tool configured that ran fairly
slowly, it was causing the command to timeout.

Since the extension isn't even reading the diff, it's better to switch
to jj log to avoid generating the diff.